Rename package to configurator2
This is to avoid confusion/problems as this is going to be a massively
breaking fork,  mostly in terms of source code,  not the syntax
of the configuration file itself.

There's a lot more I don't like about configurator,  but I think these
changes are the minimum to get me going for now.