Export essential config manipulations from Parser module
Also,  I'm seriously souring on the idea of `askLocal` and also thinking
that using the `Data.Configurator.Config` module should be discouraged,
as in semi-internal (but "less" internal than `Config.Internal`)

The reason is because forcing everything to go through the ConfigParser
applicative gives me more options of how to implement change notification
down the road.

The straightforward approach to implementing

    subscribe :: ConfigCache -> ConfigParser a -> (a -> IO ()) -> IO ()

is to rerun every subscription and then call the callback.  Of course,
a lot of things won't want to be notified unless their config _actually
changes_,  so the most straightfoward thing to do is to either put an
`Eq` constraint on `a` or punt the resposibility to the callback.

(This isn't too unreasonable, as the callback could store the last
config in an IORef or MVar,  then compare a newly incoming config
to the last one seen,  and then decide whether to send it on to another
callback.)

But...  a more efficient idea would be to track the values the `ConfigParser`
depends on,  and only re-run the parsers  (and call their associated
callbacks) whose results might actually change.  But we can't do this
if we can't track those dependencies.

Also,  the idea of tracking dependencies also has more uses than simply
optimizing the propogation of configuration changes:  it might also
make it possible to generate sample configurations,  or give sysadmins
the ability to see which parts of their configuration file doesn't
actually impact their application.

I'd at least like a fighting chance to do this,  thus the interface
becomes an important consideration...